requirements.txt
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
fake-useragent==1.5.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
playwright==1.43.0
pycparser==2.22
pyee==11.1.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wsproto==1.2.0
